Correction fichiers

+++ app/src/main/java/fr/uga/l3miage/photonum/image/ImageController.java
@@ -37,7 +37,7 @@ public class ImageController {
 
     // modification d'une image
     @PutMapping("client/{clientId}/images/{imageId}")
-    public ImageDTO updateImage(@PathVariable("clientId") int clientId, @PathVariable("imageId") int imageId, @Valid @RequestBody ImageDTO image) throws EntityNotFoundException {
+    public ImageDTO updateImage(@PathVariable("clientId") Long clientId, @PathVariable("imageId") Long imageId, @Valid @RequestBody ImageDTO image) throws EntityNotFoundException {
         try {
             Image updated = imageService.update(imageMapper.dtoToEntity(image));
             return imageMapper.entityToDTO(updated);
@@ -48,13 +48,13 @@ public class ImageController {
 
     //suppression d'une image
     @DeleteMapping("client/{clientId}/images/{imageId}")
-    public void deleteImage(@PathVariable("clientId") int clientId, @PathVariable("imageId") Long imageId) throws EntityNotFoundException, DeleteImageException {
+    public void deleteImage(@PathVariable("clientId") Long clientId, @PathVariable("imageId") Long imageId) throws EntityNotFoundException, DeleteImageException {
         try {
             imageService.delete(imageId);
         } catch(EntityNotFoundException e) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         } catch(DeleteImageException e) { // si l'image est partagée
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST); 
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
         }
     }
